babel syntax update
