+++ lib/CurrentStatus.js
@@ -1,0 +1,8 @@
+/**
+ * Document temporal statuses
+ */
+export default {
+  Archived:        0,
+  Current:         1,
+  InProgressClone: 2
+}
